거래 리플레이 버그를 수정하고 external.naver 전면 개선
- TradeService.replayTrades에서 포지션에 이전 거래가 하나도 없을 때(신규 포지션
  첫 거래, 동일 시각 거래 등) 무조건 예외를 던지던 버그 수정
- NaverStockProvider.getMarketPrices가 요청한 종목 중 일부만 응답에 없어도
  에러 대신 부분 결과를 반환하도록 변경, 대시보드는 가격 미확인 종목만 제외하고
  표시하도록 수정
- TradePreviewRequestDto에 executedAt을 추가해 과거 시점 매도 미리보기가 그
  시점 기준 매수평균단가를 쓰도록 수정
- 종목 카탈로그 메타데이터(isEtf/market)는 최초 등록 시점에만 결정되고 이후
  거래 수정으로는 갱신되지 않도록 정리

+++ backend/src/main/kotlin/kr/byeongmin/stockdaejang/domain/dashboard/dto/DashboardStockResponseDto.kt
@@ -83,7 +83,7 @@ data class DashboardStockResponseDto(
             val quantity = position.quantity
             val totalBuyAmount = position.totalBuyAmount
             val averageBuyPrice = totalBuyAmount.divideRounded(quantity)
-            val currentPrice = BigDecimal.valueOf(marketPrice.price)
+            val currentPrice = marketPrice.price
             val valuation = currentPrice.multiplyRounded(quantity)
             val unrealizedProfit = valuation.subtractRounded(totalBuyAmount)
             return DashboardStockResponseDto(
